% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_results_table.R
\name{make_results_table}
\alias{make_results_table}
\title{Read boot output files - which contain lrt statistics - and assemble into a single table}
\usage{
make_results_table(boot_run_num = 301,
  total_number_of_scans_per_trait = 400, nlrt_per_file = 10)
}
\arguments{
\item{boot_run_num}{integer to indicate which run of experimental design this is.}

\item{total_number_of_scans_per_trait}{number of scans ie number of 2d scans - for each trait file}

\item{nlrt_per_file}{positive integer specifying number of lrt statistics per boot output file}
}
\description{
Read boot output files - which contain lrt statistics - and assemble into a single table
}
