% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_results_table.R
\name{make_results_table2}
\alias{make_results_table2}
\title{Read boot output files - which contain lrt statistics - and assemble into a single table, try 2}
\usage{
make_results_table2(boot_run_num = 305)
}
\arguments{
\item{boot_run_num}{integer to indicate which run of experimental design this is.}
}
\description{
Read boot output files - which contain lrt statistics - and assemble into a single table, try 2
}
